Enum sota::datatype::event::Event

pub enum Event {
    Error(String),
    Authenticated,
    NotAuthenticated,
    UpdatesReceived(Vec<UpdateRequest>),
    UpdateAvailable(UpdateAvailable),
    NoUpdateRequests,
    FoundInstalledPackages(Vec<Package>),
    FoundSystemInfo(String),
    DownloadingUpdate(UpdateRequestId),
    DownloadComplete(DownloadComplete),
    DownloadFailed(UpdateRequestId, String),
    InstallingUpdate(UpdateRequestId),
    InstallComplete(UpdateReport),
    InstallFailed(UpdateReport),
    UpdateReportSent,
    InstalledPackagesSent,
    InstalledSoftwareSent,
    SystemInfoSent,
    InstalledSoftwareNeeded,
}

System-wide events that are broadcast to all interested parties.

Variants
Error(String)

General error event with a printable representation for debugging.

Authenticated

Authentication was successful.

NotAuthenticated

An operation failed because we are not currently authenticated.

UpdatesReceived(Vec<UpdateRequest>)

A notification from Core of pending or in-flight updates.

UpdateAvailable(UpdateAvailable)

A notification from RVI of a pending update.

NoUpdateRequests

There are no outstanding update requests.

FoundInstalledPackages(Vec<Package>)

The following packages are installed on the device.

FoundSystemInfo(String)

An update on the system information was received.

DownloadingUpdate(UpdateRequestId)

Downloading an update.

DownloadComplete(DownloadComplete)

An update was downloaded.

DownloadFailed(UpdateRequestId, String)

Downloading an update failed.

InstallingUpdate(UpdateRequestId)

Installing an update.

InstallComplete(UpdateReport)

An update was installed.

InstallFailed(UpdateReport)

The installation of an update failed.

UpdateReportSent

An update report was sent to the Core server.

InstalledPackagesSent

A list of installed packages was sent to the Core server.

InstalledSoftwareSent

A list of installed software was sent to the Core server.

SystemInfoSent

The system information was sent to the Core server.

InstalledSoftwareNeeded

A broadcast event requesting an update on externally installed software.