Module sota::datatype::package

Structs

ChunkReceived
A JSON-RPC request type to notify RVI that a new package chunk was received.

DownloadComplete
A notification from RVI to indicate the package download is complete.

DownloadStarted
A JSON-RPC request type to notify RVI that a new package download has started.

GetInstalledSoftware
A notification from RVI requesting a report on the installed software.

Package
Encodes the name and version of a specific package.

PendingUpdateRequest
A single pending update request to be installed by the client.

UpdateAvailable
A notification from RVI that a new update is available.

Enums

UpdateState
Track the transition states when installing a new package.

Type Definitions

UpdateRequestId
Encapsulate a String type as the id of a specific update request.

Version
Encapsulate a String type used to represent the Package version.