Module sota::datatype::update

Structs

DeviceReport
Bind the installation outcome report to a specific device.

InstalledFirmware
Encapsulates a single firmware installed on the device.

InstalledPackage
Encapsulates a single package installed on the device.

InstalledSoftware
An encodable list of packages and firmwares to send to RVI.

OperationResult
An encodable response of the installation outcome for a particular update ID.

UpdateReport
An encodable report of the installation outcome.

Enums

UpdateResultCode
Enumerate the possible outcomes when trying to install a package.