Enum sota::datatype::update::UpdateResultCode

pub enum UpdateResultCode {
    OK,
    ALREADY_PROCESSED,
    DEPENDENCY_FAILURE,
    VALIDATION_FAILED,
    INSTALL_FAILED,
    UPGRADE_FAILED,
    REMOVAL_FAILED,
    FLASH_FAILED,
    CREATE_PARTITION_FAILED,
    DELETE_PARTITION_FAILED,
    RESIZE_PARTITION_FAILED,
    WRITE_PARTITION_FAILED,
    PATCH_PARTITION_FAILED,
    USER_DECLINED,
    SOFTWARE_BLACKLISTED,
    DISK_FULL,
    NOT_FOUND,
    OLD_VERSION,
    INTERNAL_ERROR,
    GENERAL_ERROR,
}

Enumerate the possible outcomes when trying to install a package.

Variants
OK

Operation executed successfully

ALREADY_PROCESSED

Operation has already been processed

DEPENDENCY_FAILURE

Dependency failure during package install, upgrade, or removal

VALIDATION_FAILED

Update image integrity has been compromised

INSTALL_FAILED

Package installation failed

UPGRADE_FAILED

Package upgrade failed

REMOVAL_FAILED

Package removal failed

FLASH_FAILED

The module loader could not flash its managed module

CREATE_PARTITION_FAILED

Partition creation failed

DELETE_PARTITION_FAILED

Partition deletion failed

RESIZE_PARTITION_FAILED

Partition resize failed

WRITE_PARTITION_FAILED

Partition write failed

PATCH_PARTITION_FAILED

Partition patching failed

USER_DECLINED

User declined the update

SOFTWARE_BLACKLISTED

Software was blacklisted

DISK_FULL

Ran out of disk space

NOT_FOUND

Software package not found

OLD_VERSION

Tried to downgrade to older version

INTERNAL_ERROR

SWM Internal integrity error

GENERAL_ERROR

Other error