Module sota::datatype::update_request

Structs

ChunkReceived
A JSON-RPC request type to notify RVI that a new package chunk was received.

DownloadComplete
A notification to an external package manager that the package was downloaded.

DownloadFailed
A notification to an external package manager that the package download failed.

DownloadStarted
A JSON-RPC request type to notify RVI that a new package download has started.

Package
Encodes the name and version of a specific package.

UpdateAvailable
A notification from RVI that a new update is available.

UpdateRequest
A device update request from Core to be installed by the client.

Enums

UpdateRequestStatus
The status of an UpdateRequest from Core.

Type Definitions

UpdateRequestId
Encapsulate a String type as the id of a specific update request.