Struct sota::ota_plus::OTA

pub struct OTA<'c, 'h> {
    // some fields omitted
}

Encapsulate the client configuration and HTTP client used for over-the-air communication.

Methods

impl<'c, 'h> OTA<'c, 'h>

fn new(config: &'c Config, client: &'h Client) -> OTA<'c, 'h>

Creates a new instance for OTA communication.

fn endpoint(&self, path: &str) -> Url

Takes a path and returns a new endpoint of the format <Core server>/api/v1/vehicle_updates/<uuid>/<path>.

fn get_package_updates(&mut self) -> Result<Vec<PendingUpdateRequest>, Error>

Query the Core server to identify any new package updates available.

fn download_package_update(&mut self, id: UpdateRequestId) -> Result<PathBuf, Error>

Download a specific update from the Core server.

fn install_package_update(&mut self, id: UpdateRequestId, etx: &Sender<Event>) -> Result<UpdateReport, Error>

Download and install an update, sending Event::UpdateStateChanged notifications at each step.

fn update_installed_packages(&mut self) -> Result<(), Error>

Send a list of the currently installed packages to the Core server.

fn send_install_report(&mut self, update_report: &UpdateReport) -> Result<(), Error>

Send the outcome of a package installation to the Core server.

fn send_system_info(&mut self, info: &Json) -> Result<(), Error>

Send system information from the device to the Core server.