Struct sota::sota::Sota

pub struct Sota<'c, 'h> {
    // some fields omitted
}

Encapsulate the client configuration and HTTP client used for software-over-the-air updates.

Methods

impl<'c, 'h> Sota<'c, 'h>

fn new(config: &'c Config, client: &'h Client) -> Sota<'c, 'h>

Creates a new instance for Sota communication.

fn get_update_requests(&mut self) -> Result<Vec<UpdateRequest>, Error>

Query the Core server for any pending or in-flight package updates.

fn download_update(&mut self, id: UpdateRequestId) -> Result<DownloadComplete, Error>

Download a specific update from the Core server.

fn install_update<'t>(&mut self, token: Option<Cow<'t, AccessToken>>, id: UpdateRequestId) -> Result<UpdateReport, UpdateReport>

Install an update using the package manager.

fn send_installed_packages(&mut self, packages: &Vec<Package>) -> Result<(), Error>

Send a list of the currently installed packages to the Core server.

fn send_update_report(&mut self, update_report: &UpdateReport) -> Result<(), Error>

Send the outcome of a package update to the Core server.

fn send_system_info(&mut self, body: &str) -> Result<(), Error>

Send system information from the device to the Core server.